Added callback_url parameter to Twython to set the URL the remote site will use to return the request token.
Supports both OAuth 1.0 and 1.0a methods of setting the oauth_callback
parameter. For OAuth 1.0a support, a patched version of python-oauth2
is currently required (see
https://github.com/simplegeo/python-oauth2/pull/43).

Availability of OAuth 1.0a support in python-oauth2 is autodetected
and a suitable warning is given if the remote site requires it but
python-oauth2 doesn't support it.