Make html_for_tweet() cope if a tweet has no entities.
I'm not 100% sure if it's possible any more to get tweet data that doesn't
include entities. I've a feeling it might be, but can't find an example.

However, `html_for_tweet()` currently almost handles a tweet not having
entities, except it throws `UnboundLocalError` because the `text`
variable is declared in the wrong place. This fixes that.